refactor(test): split fuzz tests into per-suite files, add @MainActor and @Suite
