+++ PineTests/FuzzQuickOpenTests.swift
@@ -1,0 +1,117 @@
+//
+//  FuzzQuickOpenTests.swift
+//  PineTests
+//
+//  Fuzz tests for QuickOpenProvider fuzzy matching and scoring.
+//
+
+import Foundation
+import Testing
+@testable import Pine
+
+@Suite("Fuzz Quick Open Provider Tests")
+@MainActor
+struct FuzzQuickOpenProviderTests {
+
+    @Test func fuzzFuzzyScore_randomInput() {
+        var rng = SplitMix64(seed: 48)
+        let provider = QuickOpenProvider()
+
+        for _ in 0..<1000 {
+            let query: String
+            let fileName: String
+            let path: String
+
+            switch rng.next() % 5 {
+            case 0:
+                // Empty query
+                query = ""
+                fileName = FuzzGen.randomPrintable(count: 20, rng: &rng)
+                path = FuzzGen.randomPrintable(count: 50, rng: &rng)
+            case 1:
+                // Unicode query
+                query = FuzzGen.randomUnicode(count: FuzzGen.randomLength(max: 30, rng: &rng), rng: &rng)
+                fileName = FuzzGen.randomUnicode(count: 30, rng: &rng)
+                path = FuzzGen.randomUnicode(count: 60, rng: &rng)
+            case 2:
+                // Very long query
+                query = FuzzGen.randomPrintable(
+                    count: FuzzGen.randomLength(min: 100, max: 1000, rng: &rng),
+                    rng: &rng
+                )
+                fileName = FuzzGen.randomPrintable(count: 10, rng: &rng)
+                path = FuzzGen.randomPrintable(count: 20, rng: &rng)
+            case 3:
+                // Query longer than target
+                query = "abcdefghijklmnop"
+                fileName = "ab"
+                path = "ab"
+            default:
+                // Control characters
+                query = FuzzGen.randomBytes(count: FuzzGen.randomLength(max: 30, rng: &rng), rng: &rng)
+                fileName = FuzzGen.randomBytes(count: 20, rng: &rng)
+                path = FuzzGen.randomBytes(count: 40, rng: &rng)
+            }
+
+            // Must not crash
+            _ = provider.fuzzyScore(
+                queryLower: query.lowercased(),
+                fileNameLower: fileName.lowercased(),
+                pathLower: path.lowercased(),
+                pathLength: path.count
+            )
+        }
+    }
+
+    @Test func fuzzIsSubsequence_randomInput() {
+        var rng = SplitMix64(seed: 49)
+
+        for _ in 0..<1000 {
+            let query = FuzzGen.randomPrintable(
+                count: FuzzGen.randomLength(max: 50, rng: &rng),
+                rng: &rng
+            )
+            let target = FuzzGen.randomPrintable(
+                count: FuzzGen.randomLength(max: 100, rng: &rng),
+                rng: &rng
+            )
+            // Must not crash
+            _ = QuickOpenProvider.isSubsequence(query, of: target)
+        }
+    }
+
+    @Test func fuzzIsSubsequence_edgeCases() {
+        let cases: [(String, String)] = [
+            ("", ""),
+            ("", "abc"),
+            ("abc", ""),
+            ("a", "a"),
+            ("\0", "\0"),
+            ("\n\r\t", "\n\r\t"),
+            (String(repeating: "a", count: 10000), String(repeating: "a", count: 10000)),
+            ("abc", String(repeating: "x", count: 10000)),
+            ("\u{FEFF}", "\u{FEFF}text"),
+            ("\u{200B}", "a\u{200B}b"),
+        ]
+
+        for (query, target) in cases {
+            _ = QuickOpenProvider.isSubsequence(query, of: target)
+        }
+    }
+
+    @Test func fuzzFuzzyScore_emptyInputs() {
+        let provider = QuickOpenProvider()
+
+        // All empty
+        let r1 = provider.fuzzyScore(queryLower: "", fileNameLower: "", pathLower: "", pathLength: 0)
+        // Empty query is always a subsequence
+        #expect(r1 != nil)
+
+        // Empty file name but non-empty path
+        _ = provider.fuzzyScore(queryLower: "a", fileNameLower: "", pathLower: "a/b/c", pathLength: 5)
+
+        // Negative-ish path length (should not crash)
+        _ = provider.fuzzyScore(queryLower: "a", fileNameLower: "a", pathLower: "a", pathLength: 0)
+        _ = provider.fuzzyScore(queryLower: "a", fileNameLower: "a", pathLower: "a", pathLength: Int.max)
+    }
+}
